Add files via upload
Hi! This is my first file I'm sharing. I'm a graphic designer currently learning JavaScript via the bootcamp, SheCodes.

This was our 2nd week's homework where we created an interactive button that changes the h1 and emoji at the top of the page, based on user responses. 
- NO other content will change
- We created this daily weather info layout in a previous homework challenge
- Our challenge this time was ONLY to change the temperature number, city, and emoji based on whether the temperature was hot or cold (only 2 emoji changes required)

I went further by changing the emoji based on various levels of coldness AND based on whether you use a Celsius or Fahrenheit metric.

It was fun working with the various 'if' statements, and working with the script to get the results I imagined. Though I do not have perfect code and am still learning, I did try and condense the script as best as possible :) comments are welcome!